sửa note

+++ app/src/main/java/com/example/inote/models/Note.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Entity(tableName = "note")
 public class Note {
     @PrimaryKey(autoGenerate = true)
-    @NonNull private Long id;
+    @NonNull private int id;
     private int idFolder;
     private boolean isPinned;
     private String listImage;
@@ -41,11 +41,11 @@ public class Note {
         this.valueChecklist = valueChecklist;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
